Crate message_pool

Re-exports

pub use self::block_prob::*;

Modules

block_prob

test_provider

Structs

MessagePool
This is the main MessagePool struct

MpoolConfig

MpoolRpcProvider
This is the Provider implementation that will be used for the mpool RPC

MsgSet
Simple struct that contains a hashmap of messages where k: a message from address, v: a message which corresponds to that address

Enums

Error

Traits

Provider
Provider Trait. This trait will be used by the messagepool to interact with some medium in order to do the operations that are listed below that are required for the messagepool.

Functions

get_gas_perf

get_gas_reward

head_change
This function will revert and/or apply tipsets to the message pool. This function should be called every time that there is a head change in the message pool

remove
Remove a message from pending given the from address and sequence