Function message_pool::head_change

pub async fn head_change<T, '_, '_, '_, '_, '_>(
    api: &'_ RwLock<T>, 
    bls_sig_cache: &'_ RwLock<LruCache<Cid, Signature>>, 
    repub_trigger: Arc<Sender<()>>, 
    republished: &'_ RwLock<HashSet<Cid>>, 
    pending: &'_ RwLock<HashMap<Address, MsgSet>>, 
    cur_tipset: &'_ RwLock<Arc<Tipset>>, 
    revert: Vec<Tipset>, 
    apply: Vec<Tipset>
) -> Result<(), Error> where
    T: Provider + 'static, 

This function will revert and/or apply tipsets to the message pool. This function should be called every time that there is a head change in the message pool