Struct message_pool::MessagePool

pub struct MessagePool<T> {
    pub pending: Arc<RwLock<HashMap<Address, MsgSet>>>,
    pub cur_tipset: Arc<RwLock<Arc<Tipset>>>,
    pub api: Arc<RwLock<T>>,
    pub min_gas_price: BigInt,
    pub max_tx_pool_size: i64,
    pub network_name: String,
    pub network_sender: Sender<NetworkMessage>,
    pub bls_sig_cache: Arc<RwLock<LruCache<Cid, Signature>>>,
    pub sig_val_cache: Arc<RwLock<LruCache<Cid, ()>>>,
    pub republished: Arc<RwLock<HashSet<Cid>>>,
    pub repub_trigger: Sender<()>,
    pub config: MpoolConfig,
    // some fields omitted
}

This contains all necessary information needed for the message pool. Keeps track of messages to apply, as well as context needed for verifying transactions.

Fields
pending: Arc<RwLock<HashMap<Address, MsgSet>>>

A map of pending messages where the key is the address

cur_tipset: Arc<RwLock<Arc<Tipset>>>

The current tipset (a set of blocks)

api: Arc<RwLock<T>>

The underlying provider

min_gas_price: BigInt

The minimum gas price needed for executing the transaction based on number of included blocks

max_tx_pool_size: i64

This is max number of messages in the pool.

network_name: String

TODO

network_sender: Sender<NetworkMessage>

Sender half to send messages to other components

bls_sig_cache: Arc<RwLock<LruCache<Cid, Signature>>>

A cache for BLS signature keyed by Cid

sig_val_cache: Arc<RwLock<LruCache<Cid, ()>>>

A cache for BLS signature keyed by Cid

republished: Arc<RwLock<HashSet<Cid>>>

A set of republished messages identified by their Cid

repub_trigger: Sender<()>

Acts as a signal to republish messages from the republished set of messages

config: MpoolConfig

Configurable parameters of the message pool

Implementations

impl<T> MessagePool<T> where
T: Provider + Send + Sync + 'static,

pub async fn new(
api: T,
network_name: String,
network_sender: Sender<NetworkMessage>,
config: MpoolConfig
) -> Result<MessagePool<T>, Error> where
T: Provider,

Creates a new MessagePool instance.

pub async fn push(&self, msg: SignedMessage) -> Result<Cid, Error>

Push a signed message to the MessagePool. Additionally performs

pub async fn add(&self, msg: SignedMessage) -> Result<(), Error>

This is a helper to push that will help to make sure that the message fits the parameters to be pushed to the MessagePool.

pub async fn add_skip_checks(&mut self, m: SignedMessage) -> Result<(), Error>

Add a SignedMessage without doing any of the checks.

pub async fn get_sequence(&self, addr: &Address) -> Result<u64, Error>

Get the sequence for a given address, return Error if there is a failure to retrieve the respective sequence.

pub async fn push_with_sequence<V>(
&self,
addr: &Address,
cb: T
) -> Result<SignedMessage, Error> where
T: Fn(Address, u64) -> Result<SignedMessage, Error>,
V: ProofVerifier,

Adds a local message returned from the call back function with the current nonce.

pub async fn remove(
&mut self,
from: &Address,
sequence: u64,
applied: bool
) -> Result<(), Error>

Remove a message given a sequence and address from the messagepool.

pub async fn pending(&self) -> Result<(Vec<SignedMessage>, Arc<Tipset>), Error>

Return a tuple that contains a vector of all signed messages and the current tipset for self.

pub async fn pending_for(&self, a: &Address) -> Option<Vec<SignedMessage>>

Return a Vector of signed messages for a given from address. This vector will be sorted by each messsage’s sequence. If no corresponding messages found, return None result type.

pub async fn messages_for_blocks(
&self,
blks: &[BlockHeader]
) -> Result<Vec<SignedMessage>, Error>

Return Vector of signed messages given a block header for self.

pub fn estimate_gas_premium(
&self,
nblocksincl: u64,
_sender: Address,
_gas_limit: u64,
_tsk: TipsetKeys
) -> Result<BigInt, Error>

Return gas price estimate this has been translated from lotus, a more smart implementation will most likely need to be implemented.

pub async fn load_local(&mut self) -> Result<(), Error>

Loads local messages to the message pool to be applied.

pub async fn clear(&mut self, local: bool)

If local = true, the local messages will be removed as well as pending messages. If local = false, pending messages will be removed while retaining local messages.

pub fn get_config(&self) -> &MpoolConfig

pub fn set_config<DB: Store>(
&mut self,
db: &DB,
cfg: MpoolConfig
) -> Result<(), Error>

impl<T> MessagePool<T> where
T: Provider + Send + Sync + 'static,

pub async fn select_messages(
&self,
ts: &Tipset,
tq: f64
) -> Result<Vec<SignedMessage>, Error>

Forest employs a sophisticated algorithm for selecting messages for inclusion from the pool, given the ticket quality of a miner. This method selects messages for including in a block.