Trait message_pool::Provider

pub trait Provider {
    fn subscribe_head_changes(&mut self) -> Subscriber<HeadChange>;

    fn get_heaviest_tipset(&mut self) -> Option<Tipset>;

    fn put_message(&self, msg: &SignedMessage) -> Result<Cid, Error>;

    fn state_get_actor(
        &self, 
        addr: &Address, 
        ts: &Tipset
    ) -> Result<ActorState, Error>;

    fn messages_for_block(
        &self, 
        h: &BlockHeader
    ) -> Result<(Vec<UnsignedMessage>, Vec<SignedMessage>), Error>;

    fn messages_for_tipset(
        &self, 
        h: &Tipset
    ) -> Result<Vec<UnsignedMessage>, Error>;

    fn load_tipset(&self, tsk: &TipsetKeys) -> Result<Tipset, Error>;
}

Provider Trait. This trait will be used by the messagepool to interact with some medium in order to do the operations that are listed below that are required for the messagepool.

Required methods

fn subscribe_head_changes(&mut self) -> Subscriber<HeadChange>

Update Mpool's cur_tipset whenever there is a chnge to the provider

fn get_heaviest_tipset(&mut self) -> Option<Tipset>

Get the heaviest Tipset in the provider

fn put_message(&self, msg: &SignedMessage) -> Result<Cid, Error>

Add a message to the MpoolProvider, return either Cid or Error depending on successful put

fn state_get_actor(
&self,
addr: &Address,
ts: &Tipset
) -> Result<ActorState, Error>

Return state actor for given address given the tipset that the a temp StateTree will be rooted at. Return ActorState or Error depending on whether or not ActorState is found

fn messages_for_block(
&self,
h: &BlockHeader
) -> Result<(Vec<UnsignedMessage>, Vec<SignedMessage>), Error>

Return the signed messages for given blockheader

fn messages_for_tipset(&self, h: &Tipset) -> Result<Vec<UnsignedMessage>, Error>

Return all messages for a tipset

fn load_tipset(&self, tsk: &TipsetKeys) -> Result<Tipset, Error>

Return a tipset given the tipset keys from the ChainStore

Loading content...