Crate message_pool

Modules

test_provider
Contains mock implementations for testing internal MessagePool APIs

Structs

MessagePool
This contains all necessary information needed for the message pool. Keeps track of messages to apply, as well as context needed for verifying transactions.

MpoolConfig
Config available for the MessagePool.

MpoolRpcProvider
This is the default Provider implementation that will be used for the mpool RPC.

Enums

Error
MessagePool error.

Traits

Provider
Provider Trait. This trait will be used by the messagepool to interact with some medium in order to do the operations that are listed below that are required for the messagepool.

Functions

block_probabilities

head_change
This function will revert and/or apply tipsets to the message pool. This function should be called every time that there is a head change in the message pool.