Trait message_pool::Provider

pub trait Provider {
    #[must_use]
    fn subscribe_head_changes<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Subscriber<HeadChange>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_heaviest_tipset<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Option<Arc<Tipset>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn put_message(&self, msg: &ChainMessage) -> Result<Cid, Error>;

    fn get_actor_after(
        &self, 
        addr: &Address, 
        ts: &Tipset
    ) -> Result<ActorState, Error>;

    fn messages_for_block(
        &self, 
        h: &BlockHeader
    ) -> Result<(Vec<UnsignedMessage>, Vec<SignedMessage>), Error>;

    #[must_use]
    fn state_account_key<'life0, 'life1, 'life2, 'async_trait, V>(
        &'life0 self, 
        addr: &'life1 Address, 
        ts: &'life2 Arc<Tipset>
    ) -> Pin<Box<dyn Future<Output = Result<Address, Error>> + Send + 'async_trait>>
    where
        V: ProofVerifier,
        V: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn messages_for_tipset(
        &self, 
        h: &Tipset
    ) -> Result<Vec<ChainMessage>, Error>;

    #[must_use]
    fn load_tipset<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        tsk: &'life1 TipsetKeys
    ) -> Pin<Box<dyn Future<Output = Result<Arc<Tipset>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn chain_compute_base_fee(&self, ts: &Tipset) -> Result<BigInt, Error>;
}

Provider Trait. This trait will be used by the messagepool to interact with some medium in order to do the operations that are listed below that are required for the messagepool.

Required methods

#[must_use]
fn subscribe_head_changes<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Subscriber<HeadChange>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Update Mpool’s cur_tipset whenever there is a chnge to the provider

#[must_use]
fn get_heaviest_tipset<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Option<Arc<Tipset>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get the heaviest Tipset in the provider

fn put_message(&self, msg: &ChainMessage) -> Result<Cid, Error>

Add a message to the MpoolProvider, return either Cid or Error depending on successful put

fn get_actor_after(
&self,
addr: &Address,
ts: &Tipset
) -> Result<ActorState, Error>

Return state actor for given address given the tipset that the a temp StateTree will be rooted at. Return ActorState or Error depending on whether or not ActorState is found

fn messages_for_block(
&self,
h: &BlockHeader
) -> Result<(Vec<UnsignedMessage>, Vec<SignedMessage>), Error>

Return the signed messages for given blockheader

#[must_use]
fn state_account_key<'life0, 'life1, 'life2, 'async_trait, V>(
&'life0 self,
addr: &'life1 Address,
ts: &'life2 Arc<Tipset>
) -> Pin<Box<dyn Future<Output = Result<Address, Error>> + Send + 'async_trait>> where
V: ProofVerifier,
V: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Resolves to the key address

fn messages_for_tipset(&self, h: &Tipset) -> Result<Vec<ChainMessage>, Error>

Return all messages for a tipset

#[must_use]
fn load_tipset<'life0, 'life1, 'async_trait>(
&'life0 self,
tsk: &'life1 TipsetKeys
) -> Pin<Box<dyn Future<Output = Result<Arc<Tipset>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Return a tipset given the tipset keys from the ChainStore

fn chain_compute_base_fee(&self, ts: &Tipset) -> Result<BigInt, Error>

Computes the base fee

Loading content...